backend/requirements.txt
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
google-ai-generativelanguage==0.1.0
google-api-core==2.25.0
google-auth==2.40.3
google-generativeai==0.1.0rc1
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.62.3
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-dotenv==1.0.1
rdflib==7.1.4
requests==2.32.4
rsa==4.9.1
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
